@@ -2,12 +2,13 @@ import { Component, Inject, OnInit, ViewChild } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { MatSort } from '@angular/material/sort';
+import { MatSort, Sort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { TeachingPeriodBreak } from 'src/app/api/models/teaching-period';
 import { TeachingPeriod } from 'src/app/api/models/teaching-period';
 import { TeachingPeriodBreakService } from 'src/app/api/services/teaching-period-break.service';
 import { TeachingPeriodService } from 'src/app/api/services/teaching-period.service';
+import { TeachingPeriodUnitImportService } from '../teaching-period-unit-import/teaching-period-unit-import.dialog';
 
 @Component({
   selector: 'f-teaching-period-list',
@@ -20,8 +21,13 @@ export class TeachingPeriodListComponent implements OnInit {
 
   public dataSource = new MatTableDataSource<TeachingPeriod>();
 
-  displayedColumns: string[] = ['active', 'name', 'startDate', 'endDate', 'activeUntil'];
-  constructor(private teachingPeriodsService: TeachingPeriodService, public dialog: MatDialog) {}
+  displayedColumns: string[] = ['active', 'name', 'startDate', 'endDate', 'activeUntil', 'actions'];
+
+  constructor(
+    private teachingPeriodsService: TeachingPeriodService,
+    public dialog: MatDialog,
+    public teachingPeriodUnitImportService: TeachingPeriodUnitImportService,
+  ) {}
 
   ngOnInit(): void {
     // update the Teaching Periods
@@ -35,6 +41,10 @@ export class TeachingPeriodListComponent implements OnInit {
     });
   }
 
+  importUnits(teachingPeriod: TeachingPeriod) {
+    this.teachingPeriodUnitImportService.openImportUnitsDialog(teachingPeriod);
+  }
+
   addTeachingPeriod() {
     this.dialog.open(NewTeachingPeriodDialogComponent, {
       data: {},
@@ -45,6 +55,39 @@ export class TeachingPeriodListComponent implements OnInit {
     this.teachingPeriodsService.get(selectedTeachingPeriod.id).subscribe((teachingPeriod) => {
       this.dialog.open(NewTeachingPeriodDialogComponent, { data: { teachingPeriod: teachingPeriod } });
     });
+  }
+
+  /**
+   * Function used by implemented sortTableData to determine the order
+   * of values within the EntityForm once sorting has been triggered.
+   *
+   * @param aValue value to be compared against bValue.
+   * @param bValue value to be compared against aValue.
+   *
+   * @returns truthy comparison between aValue and bValue.
+   */
+  protected sortCompare(aValue: number | string, bValue: number | string, isAsc: boolean) {
+    return (aValue < bValue ? -1 : 1) * (isAsc ? 1 : -1);
+  }
+
+  // Sorting function to sort data when sort
+  // event is triggered
+  sortTableData(sort: Sort) {
+    if (!sort.active || sort.direction === '') {
+      return;
+    }
+    switch (sort.active) {
+      case 'active':
+      case 'name':
+      case 'startDate':
+      case 'endDate':
+      case 'activeUntil':
+        this.dataSource.data = this.dataSource.data.sort((a, b) => {
+          const isAsc = sort.direction === 'asc';
+          return this.sortCompare(a[sort.active], b[sort.active], isAsc);
+        });
+        return;
+    }
   }
 }
 
