@@ -21,7 +21,7 @@ export class TaskDefinition extends Entity {
   targetDate: Date;
   dueDate: Date;
   startDate: Date;
-  uploadRequirements: UploadRequirement[];
+  uploadRequirements: UploadRequirement[] = [];
   tutorialStream: TutorialStream = null;
   plagiarismChecks: SimilarityCheck[] = [];
   plagiarismReportUrl: string;
@@ -31,6 +31,12 @@ export class TaskDefinition extends Entity {
   groupSet: GroupSet = null;
   hasTaskSheet: boolean;
   hasTaskResources: boolean;
+  scormEnabled: boolean;
+  hasScormData: boolean;
+  scormAllowReview: boolean;
+  scormBypassTest: boolean;
+  scormTimeDelayEnabled: boolean;
+  scormAttemptLimit: number = 0;
   hasTaskAssessmentResources: boolean;
   isGraded: boolean;
   maxQualityPts: number;
@@ -152,6 +158,20 @@ export class TaskDefinition extends Entity {
     }`;
   }
 
+  public getScormDataUrl(asAttachment: boolean = false) {
+    const constants = AppInjector.get(DoubtfireConstants);
+    return `${constants.API_URL}/units/${this.unit.id}/task_definitions/${this.id}/scorm_data.json${
+      asAttachment ? '?as_attachment=true' : ''
+    }`;
+  }
+
+  /**
+   * Open the SCORM test in a new tab - using preview mode.
+   */
+  public previewScormTest(): void {
+    window.open(`#/task_def_id/${this.id}/preview-scorm`, '_blank');
+  }
+
   public get targetGradeText(): string {
     return Grade.GRADES[this.targetGrade];
   }
@@ -176,6 +196,12 @@ export class TaskDefinition extends Entity {
     }/task_resources`;
   }
 
+  public get scormDataUploadUrl(): string {
+    return `${AppInjector.get(DoubtfireConstants).API_URL}/units/${this.unit.id}/task_definitions/${
+      this.id
+    }/scorm_data`;
+  }
+
   public get taskAssessmentResourcesUploadUrl(): string {
     return `${AppInjector.get(DoubtfireConstants).API_URL}/units/${this.unit.id}/task_definitions/${
       this.id
@@ -196,6 +222,11 @@ export class TaskDefinition extends Entity {
   public deleteTaskResources(): Observable<any> {
     const httpClient = AppInjector.get(HttpClient);
     return httpClient.delete(this.taskResourcesUploadUrl).pipe(tap(() => (this.hasTaskResources = false)));
+  }
+
+  public deleteScormData(): Observable<any> {
+    const httpClient = AppInjector.get(HttpClient);
+    return httpClient.delete(this.scormDataUploadUrl).pipe(tap(() => (this.hasScormData = false)));
   }
 
   public deleteTaskAssessmentResources(): Observable<any> {
