@@ -50,15 +50,15 @@ export class UnitService extends CachedEntityService<Unit> {
         toEntityFn: (data: object, jsonKey: string, entity: Unit) => {
           const unitRoleService = AppInjector.get(UnitRoleService);
           unitRoleService.cache.get(data[jsonKey]);
-        }
+        },
       },
       {
-        keys: 'staff',
+        keys: 'unitRoles',
         toEntityOp: (data, key, entity) => {
           const unitRoleService = AppInjector.get(UnitRoleService);
           // Add staff
           entity.staffCache.clear();
-          data[key].forEach(staff => {
+          data[key]?.forEach(staff => {
             entity.staffCache.add(unitRoleService.buildInstance(staff));
           });
         }
@@ -66,7 +66,9 @@ export class UnitService extends CachedEntityService<Unit> {
       {
         keys: ['mainConvenor', 'main_convenor_id'],
         toEntityFn: (data, key, entity) => {
-          return entity.staffCache.get(data[key]);
+          let result = entity.staffCache.get(data[key]);
+          entity.mainConvenorUser = result?.user;
+          return result;
         },
         toJsonFn: (unit: Unit, key: string) => {
           return unit.mainConvenor?.id;
@@ -133,7 +135,7 @@ export class UnitService extends CachedEntityService<Unit> {
       {
         keys: 'ilos',
         toEntityOp: (data: object, key: string, unit: Unit) => {
-          data[key].forEach(ilo => {
+          data[key]?.forEach(ilo => {
             unit.learningOutcomesCache.getOrCreate(ilo['id'], this.learningOutcomeService, ilo);
           });
         }
@@ -160,7 +162,7 @@ export class UnitService extends CachedEntityService<Unit> {
       {
         keys: 'groupSets',
         toEntityOp: (data, key, unit) => {
-          data[key].forEach((groupSetJson: object) => {
+          data[key]?.forEach((groupSetJson: object) => {
             unit.groupSetsCache.add(this.groupSetService.buildInstance(groupSetJson, {constructorParams: unit}));
           });
         }
@@ -168,7 +170,7 @@ export class UnitService extends CachedEntityService<Unit> {
       {
         keys: 'groups',
         toEntityOp: (data, key, unit) => {
-          data[key].forEach((groupJson: object) => {
+          data[key]?.forEach((groupJson: object) => {
             const group = this.groupService.buildInstance(groupJson, {constructorParams: unit});
             group.groupSet.groupsCache.add(group);
           });
