@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { EditProfileFormComponent } from './edit-profile-form.component';
+import {EditProfileFormComponent} from './edit-profile-form.component';
 
 describe('EditProfileComponent', () => {
   let component: EditProfileFormComponent;
