@@ -1,4 +1,4 @@
 let API_URL: string;
 
-API_URL = `${window.location.protocol}//${window.location.hostname}:3000/api`;
+API_URL = `${window.location.protocol}//${window.location.hostname}:4200/api`;
 export default API_URL;
